pyproject.toml
[tool.poetry]
name = "socket.io_base"
version = "0.1.0"
description = "SOCET_TRAINIE"
authors = ["BotArchitect1 <pavliksha@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
loguru = "^0.7.2"
python-socketio = "^5.11.1"
uvicorn = "^0.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
